Broke out AutoSize = false tests to new file
